verifying correctness of trades script
